+++ src/between.cpp
@@ -18,6 +18,12 @@ LogicalVector between(NumericVector x, double left, double right) {
   int n = x.size();
   LogicalVector out = no_init(n);
 
+  if (NumericVector::is_na(left) || NumericVector::is_na(right)) {
+    for (int i = 0; i < n; ++i)
+      out[i] = NA_REAL;
+    return out;
+  }
+
   for (int i = 0; i < n; ++i) {
     if (NumericVector::is_na(x[i])) {
       out[i] = NA_REAL;