+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.5.0.9000
 
+* `between()` returns NA if `left` or `right` is `NA` (fixes #2562).
+
 * `substr()` is now translated to SQL, correcting for the difference
   in the third argument. In R, it's the position of the last character,
   in SQL it's the length of the string (#2536).